[Fix] Fixes from @greymag pt. 2
